==> isoforms/R/00_seurat_object_melanoma.R <==
# create Seurat object of melanoma patient TIL4 and PBMC7

library(dplyr)
library(ggplot2)
library(Seurat)

seurat.data <- Read10X(data.dir = paste0("./data/TCR/Pool80_4-9/"))
rownames(x = seurat.data[["Antibody Capture"]]) <- gsub(pattern = "*_CITEseq", replacement = "", rownames(seurat.data[["Antibody Capture"]]))
TIL4 <- CreateSeuratObject(counts = seurat.data$`Gene Expression`, project = "TIL4", min.cells = 3, min.features = 200)
TIL4[["ADT"]] <- CreateAssayObject(seurat.data$`Antibody Capture`[, colnames(x = TIL4)])
TIL4 <- NormalizeData(TIL4, assay = "ADT", normalization.method = "CLR")
TIL4[["percent.mt"]] <- PercentageFeatureSet(TIL4, pattern = "^MT-")
TIL4 <- subset(TIL4, subset = nFeature_RNA > 200 & nFeature_RNA < 2500 & percent.mt < 20)
TIL4 <- NormalizeData(TIL4)
TIL4 <- FindVariableFeatures(TIL4)
TIL4 <- ScaleData(TIL4)
TIL4 <- RunPCA(TIL4)
TIL4 <- RunUMAP(TIL4, dims = 1:30)
TIL4 <- RenameCells(TIL4, new.names = gsub(colnames(TIL4), pattern = "-1", replacement = ""))
saveRDS("./data/TCR/objects/20230104_TIL4.rds", object = TIL4)

seurat.data <- Read10X(data.dir = paste0("./data/TCR/Pool80_4-12/"))
rownames(x = seurat.data[["Antibody Capture"]]) <- gsub(pattern = "*_CITEseq", replacement = "", rownames(seurat.data[["Antibody Capture"]]))
PBMC7 <- CreateSeuratObject(counts = seurat.data$`Gene Expression`, project = "PBMC7", min.cells = 3, min.features = 200)
PBMC7[["ADT"]] <- CreateAssayObject(seurat.data$`Antibody Capture`[, colnames(x = PBMC7)])
PBMC7 <- NormalizeData(PBMC7, assay = "ADT", normalization.method = "CLR")
PBMC7[["percent.mt"]] <- PercentageFeatureSet(PBMC7, pattern = "^MT-")
PBMC7 <- subset(PBMC7, subset = nFeature_RNA > 200 & nFeature_RNA < 2500 & percent.mt < 20)
PBMC7 <- NormalizeData(PBMC7)
PBMC7 <- FindVariableFeatures(PBMC7)
PBMC7 <- ScaleData(PBMC7)
PBMC7 <- RunPCA(PBMC7)
PBMC7 <- RunUMAP(PBMC7, dims = 1:30)
PBMC7 <- RenameCells(PBMC7, new.names = gsub(colnames(PBMC7), pattern = "-1", replacement = ""))
saveRDS("./data/TCR/objects/20230104_PBMC7.rds", object = PBMC7)
